Title: 세 수의 합과 평균 구하기, Time: 143ms, Memory: 10MB, Status: Passed - CodeTree

+++ 241202/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -10,8 +10,8 @@ public class Main {
         int c = sc.nextInt();
 
         int sum = a + b + c;
-        double avg = (double)(sum) / 3;
+        int avg = sum / 3;
 
-        System.out.printf("%d\n%.0f", sum, avg);
+        System.out.printf("%d\n%d", sum, avg);
     }
 }
